requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pytest
httpx
pytest-asyncio
pydantic
sqlmodel